refactor: centralize db access
